# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moon Paris Perfume Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  fullName: string;
  phone?: string | null;
  email?: string | null;
  governorate?: string | null;
  district?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface RegisterRequest {
  fullName: string;
  phone: string;
  email?: string | null;
  password?: string | null;
  governorate: string;
  district: string;
  otpCode: string;
}

export interface LoginRequest {
  phone?: string | null;
  email?: string | null;
  password?: string | null;
  otpCode?: string | null;
}

export interface SendOtpRequest {
  phone: string;
}

export interface VerifyOtpRequest {
  phone: string;
  code: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string | null;
  descriptionAr?: string | null;
  price: number;
  originalPrice?: number | null;
  imageUrl?: string | null;
  images: string[];
  category: string;
  brand?: string | null;
  volume?: string | null;
  inStock: boolean;
  stockQuantity: number;
  featured: boolean;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  nameAr: string;
  description?: string | null;
  descriptionAr?: string | null;
  price: number;
  originalPrice?: number | null;
  imageUrl?: string | null;
  images?: string[];
  category: string;
  brand?: string | null;
  volume?: string | null;
  inStock: boolean;
  stockQuantity: number;
  featured: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productNameAr: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  userId?: number | null;
  guestName?: string | null;
  guestPhone?: string | null;
  guestGovernorate?: string | null;
  guestDistrict?: string | null;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  notes?: string | null;
  createdAt: string;
}

export type CreateOrderRequestItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateOrderRequest {
  guestName?: string | null;
  guestPhone?: string | null;
  guestGovernorate?: string | null;
  guestDistrict?: string | null;
  items: CreateOrderRequestItemsItem[];
  notes?: string | null;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalUsers: number;
  pendingOrders: number;
}

export interface SiteSettings {
  siteName: string;
  heroTitle: string;
  heroSubtitle: string;
  heroImageUrl?: string | null;
  contactPhone?: string | null;
  contactEmail?: string | null;
  aboutText?: string | null;
}

export type GetProductsParams = {
  category?: string;
  search?: string;
};