# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Moon Paris Perfume Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP via WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders (admin or user's own)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /site-settings:
    get:
      operationId: getSiteSettings
      tags: [admin]
      summary: Get site settings
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

    put:
      operationId: updateSiteSettings
      tags: [admin]
      summary: Update site settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        governorate:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
      required:
        - id
        - fullName
        - role
        - createdAt
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        governorate:
          type: string
        district:
          type: string
        otpCode:
          type: string
      required:
        - fullName
        - phone
        - governorate
        - district
        - otpCode
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        otpCode:
          type: string
          nullable: true
      required: []
    SendOtpRequest:
      type: object
      properties:
        phone:
          type: string
      required:
        - phone
    VerifyOtpRequest:
      type: object
      properties:
        phone:
          type: string
        code:
          type: string
      required:
        - phone
        - code
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        category:
          type: string
        brand:
          type: string
          nullable: true
        volume:
          type: string
          nullable: true
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        featured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - nameAr
        - price
        - category
        - inStock
        - stockQuantity
        - featured
        - images
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        category:
          type: string
        brand:
          type: string
          nullable: true
        volume:
          type: string
          nullable: true
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        featured:
          type: boolean
      required:
        - name
        - nameAr
        - price
        - category
        - inStock
        - stockQuantity
        - featured
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        guestName:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        guestGovernorate:
          type: string
          nullable: true
        guestDistrict:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - items
        - totalAmount
        - status
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productNameAr:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - productNameAr
        - quantity
        - price
    CreateOrderRequest:
      type: object
      properties:
        guestName:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        guestGovernorate:
          type: string
          nullable: true
        guestDistrict:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        notes:
          type: string
          nullable: true
      required:
        - items
...
    SiteSettings:
      type: object
      properties:
        siteName:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        heroImageUrl:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        aboutText:
          type: string
          nullable: true
      required:
        - siteName
        - heroTitle
        - heroSubtitle