# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moon Paris Perfume Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestItemsItem } from "./createOrderRequestItemsItem";

export interface CreateOrderRequest {
  guestName?: string | null;
  guestPhone?: string | null;
  guestGovernorate?: string | null;
  guestDistrict?: string | null;
  items: CreateOrderRequestItemsItem[];
  notes?: string | null;
}