# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moon Paris Perfume Store API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  phone?: string | null;
  email?: string | null;
  password?: string | null;
  otpCode?: string | null;
}