# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moon Paris Perfume Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string | null;
  descriptionAr?: string | null;
  price: number;
  originalPrice?: number | null;
  imageUrl?: string | null;
  images: string[];
  category: string;
  brand?: string | null;
  volume?: string | null;
  inStock: boolean;
  stockQuantity: number;
  featured: boolean;
  createdAt: string;
}