# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moon Paris Perfume Store API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  fullName: string;
  phone: string;
  email?: string | null;
  password?: string | null;
  governorate: string;
  district: string;
  otpCode: string;
}